pyproject.toml
[project]
name = "100py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "briefcase>=0.3.26",
    "pillow>=12.0.0",
    "raylib>=5.5.0.3",
    "toga>=0.5.2",
]



